+++ R/setparms_LWFB90.R
@@ -29,8 +29,11 @@
 #'  qffc         \tab Quickflow fraction of infiltrating water at field capacity, for value 0 there is no quickflow (bypass or surface) unless soil profile surface becomes saturated. Default: 0  \tab -                  \tab Flow                   \cr
 #'  qfpar        \tab Quickflow shape parameter. Default: 1 \tab -                  \tab Flow                   \cr
 #'  qlayer       \tab Number of layers which are considered for generation of surface or source area flow. Default: 0\tab -                  \tab Flow                   \cr
+#'  water_table_depth \tab Depth of the water table for a constant head boundary. -9999 means gravitational flow boundary. Default: -9999\tab m                  \tab Flow                   \cr
 #'  gwatini      \tab Initial value of groundwater storage. Default: 0\tab mm                  \tab Initial                \cr
 #'  snowini      \tab Initial value of water content of snow pack. Default: 0\tab mm \tab Initial                \cr
+#'  snowlqini    \tab Initial value of liquid water content of snow pack. Default: 0\tab mm \tab Initial                \cr
+#'  snowccini    \tab Initial value of cold content of snow pack (positive). Default: 0\tab  MJ m-2 \tab Initial                \cr
 #'  intrainini   \tab Initial value of intercepted rain. Default: 0\tab mm                  \tab Initial                \cr
 #'  intsnowini   \tab Initial value of intercepted snow. Default: 0\tab -                  \tab Initial                \cr
 #'  psiini       \tab Initial pressure head of soil layers. May have the same length as row.names(soil). Default: -6.3 \tab kPa                \tab Initial  \cr
@@ -202,6 +205,7 @@ set_paramLWFB90 <- function(...) {
     gsp = 0,
     ilayer = 1,
     qlayer = 0,
+    water_table_depth = -9999,
     z0s = 0.001,
     rstemp = -0.5,
     melfac = 1.5,
@@ -239,6 +243,8 @@ set_paramLWFB90 <- function(...) {
     intsnowini = 0,
     gwatini = 0,
     snowini = 0,
+    snowccini = 0,
+    snowlqini = 0,
     psiini = -6.3
   )